+++ apps/stripe/graphql.config.ts
@@ -52,6 +52,7 @@ const config: IGraphQLConfig = {
                   WeightScalar: "number",
                   Day: "string",
                   Hour: "number",
+                  PositiveInt: "number",
                 },
               },
               plugins: [