+++ apps/stripe/src/__tests__/mocks/constants.ts
@@ -1,3 +1,5 @@
+import { SaleorSchemaVersion } from "@saleor/app-sdk/types";
+
 import { mockedSaleorApiUrl } from "@/__tests__/mocks/saleor-api-url";
 import { WebhookParams } from "@/app/api/webhooks/stripe/webhook-params";
 import { SaleorMoney } from "@/modules/saleor/saleor-money";
@@ -6,6 +8,10 @@ import { createSaleorTransactionId } from "@/modules/saleor/saleor-transaction-i
 export const mockedSaleorChannelId = "Q2hhbm5lbDox";
 export const mockedConfigurationId = "81f323bd-91e2-4838-ab6e-5affd81ffc3b";
 export const mockedSaleorAppId = "saleor-app-id";
+export const mockedSaleorSchemaVersionSupportingPaymentMethodDetails: SaleorSchemaVersion = [3, 22];
+export const mockedSaleorSchemaVersionNotSupportingPaymentMethodDetails: SaleorSchemaVersion = [
+  3, 21,
+];
 export const mockedAppToken = "XXXYYYZZZ";
 export const mockAppUrlBase = "https://my-app.saleor.app";
 export const mockAdyenWebhookUrl = `${mockAppUrlBase}?${new URLSearchParams({