+++ apps/stripe/src/__tests__/mocks/mocked-recorded-transaction.ts
@@ -1,4 +1,9 @@
-import { mockedSaleorTransactionId } from "@/__tests__/mocks/constants";
+import { SaleorSchemaVersion } from "@saleor/app-sdk/types";
+
+import {
+  mockedSaleorSchemaVersionSupportingPaymentMethodDetails,
+  mockedSaleorTransactionId,
+} from "@/__tests__/mocks/constants";
 import { mockedStripePaymentIntentId } from "@/__tests__/mocks/mocked-stripe-payment-intent-id";
 import {
   createResolvedTransactionFlow,
@@ -19,6 +24,7 @@ type Params = {
   selectedPaymentMethod?: PaymentMethod["type"];
   resolvedTransactionFlow?: ResolvedTransactionFlow;
   saleorTransactionFlow?: SaleorTransationFlow;
+  saleorSchemaVersion?: SaleorSchemaVersion;
 };
 
 export const getMockedRecordedTransaction = (params?: Params): RecordedTransaction => {
@@ -28,6 +34,7 @@ export const getMockedRecordedTransaction = (params?: Params): RecordedTransacti
     saleorTransactionFlow: createSaleorTransactionFlow("CHARGE"),
     resolvedTransactionFlow: createResolvedTransactionFlow("CHARGE"),
     selectedPaymentMethod: "card",
+    saleorSchemaVersion: mockedSaleorSchemaVersionSupportingPaymentMethodDetails,
     ...(params ?? {}),
   } satisfies Params;
 
@@ -37,5 +44,6 @@ export const getMockedRecordedTransaction = (params?: Params): RecordedTransacti
     saleorTransactionFlow: finalParams.saleorTransactionFlow,
     resolvedTransactionFlow: finalParams.resolvedTransactionFlow,
     selectedPaymentMethod: finalParams.selectedPaymentMethod,
+    saleorSchemaVersion: finalParams.saleorSchemaVersion,
   });
 };