+++ apps/stripe/src/modules/transactions-recording/repositories/dynamodb/dynamodb-transaction-recorder-repo.test.ts
@@ -93,6 +93,7 @@ describe("DynamoDBTransactionRecorderRepo", () => {
             saleorTransactionFlow: "CHARGE",
             resolvedTransactionFlow: "CHARGE",
             selectedPaymentMethod: "card",
+            saleorSchemaVersion: { major: 3, minor: 22 },
             createdAt: "2023-01-01T00:00:00.000Z",
             modifiedAt: "2023-01-01T00:00:00.000Z",
             saleorApiUrl: mockedSaleorApiUrl,
@@ -116,6 +117,10 @@ describe("DynamoDBTransactionRecorderRepo", () => {
       expect(result._unsafeUnwrap()).toMatchInlineSnapshot(`
         RecordedTransaction {
           "resolvedTransactionFlow": "CHARGE",
+          "saleorSchemaVersion": [
+            3,
+            22,
+          ],
           "saleorTransactionFlow": "CHARGE",
           "saleorTransactionId": "mocked-transaction-id",
           "selectedPaymentMethod": "card",