+++ apps/stripe/src/modules/transactions-recording/repositories/dynamodb/recorded-transaction-db-model.ts
@@ -1,4 +1,4 @@
-import { Entity, string } from "dynamodb-toolbox";
+import { Entity, map, number, string } from "dynamodb-toolbox";
 import { item } from "dynamodb-toolbox/schema/item";
 
 import { DynamoMainTable, dynamoMainTable } from "@/modules/dynamodb/dynamo-main-table";
@@ -24,6 +24,10 @@ const Schema = item({
   saleorTransactionFlow: string(),
   resolvedTransactionFlow: string(),
   selectedPaymentMethod: string(),
+  saleorSchemaVersion: map({
+    major: number(),
+    minor: number(),
+  }),
 });
 
 const createEntity = (table: DynamoMainTable) => {