+++ packages/shared/src/saleor-version-compatibility-validator.ts
@@ -1,7 +1,12 @@
+import { SaleorSchemaVersion } from "@saleor/app-sdk/types";
 import semver from "semver";
 
 export class SaleorVersionCompatibilityValidator {
-  constructor(private appRequiredVersion: string) {}
+  private appRequiredVersion: string;
+
+  constructor(appRequiredVersion: string) {
+    this.appRequiredVersion = appRequiredVersion;
+  }
 
   isValid(saleorVersion: string) {
     return semver.satisfies(saleorVersion, this.appRequiredVersion, {
@@ -15,5 +20,13 @@ export class SaleorVersionCompatibilityValidator {
         `Your Saleor version (${saleorVersion}) doesn't match App's required version (semver: ${this.appRequiredVersion})`,
       );
     }
+  }
+
+  isSaleorCompatible(saleorSchemaVersion: SaleorSchemaVersion) {
+    const [major, minor] = saleorSchemaVersion;
+
+    return semver.satisfies(`${major}.${minor}.0`, this.appRequiredVersion, {
+      includePrerelease: true,
+    });
   }
 }